refactor: transfer all images from assets/ to assets/images, fix path to images
