+++ src/assets/mock/index.ts
@@ -1,3 +1,4 @@
 export { props as SubjectList } from './TeacherList';
 export { props as TeacherList } from './SubjectList';
 export type { PropsType, LinkItem } from './mockType';
+export { news as newsList } from './newsList';
