+++ src/assets/mock/newsList.ts
@@ -1,0 +1,38 @@
+import { NewsItem } from './mockType';
+
+export const news: NewsItem[] = [
+  {
+    id: 0,
+    title: 'Заголовок лайфхака',
+    comments: 35,
+    likes: 65,
+    favorites: 13,
+    text: `Давно выяснено, что при оценке дизайна и композиции читаемый текст
+      мешает сосредоточиться. Lorem Ipsum используют потому, что тот
+      обеспечивает более или менее стандартное заполнение шаблона, а
+      также реальное распределение букв и пробелов в абзацах, которое не
+      получается при простой дубликации "Здесь ваш текст.. Здесь ваш
+      текст.. Здесь ваш текст.." Многие программы электронной вёрстки и
+      редакторы HTML используют Lorem Ipsum в качестве текста по
+      умолчанию, так что поиск по ключевым словам "lorem ipsum" сразу
+      показывает, как много веб-страниц всё ещё дожидаются своего
+      настоящего рождения. За прошедшие годы текст Lorem Ipsum получил
+      много версий. Некоторые версии появились по ошибке, некоторые -
+      намеренно (например, юмористические варианты).`,
+    published: '12.12.2019',
+  },
+  {
+    id: 1,
+    title: 'Второй заголовок лайфхака',
+    comments: 10,
+    likes: 21,
+    favorites: 7,
+    text: `Давно выяснено, что при оценке дизайна и композиции читаемый текст
+      мешает сосредоточиться. Lorem Ipsum используют потому, что тот
+      обеспечивает более или менее стандартное заполнение шаблона, а
+      также реальное распределение букв и пробелов в абзацах, которое не
+      получается при простой дубликации "Здесь ваш текст.. Здесь ваш
+      текст.. Здесь ваш текст.."`,
+    published: '11.12.2019',
+  },
+];
